refactor(routes): extract process-control routes to agents-process.ts
Moves auto-restart, context-guard (per-agent + fleet /api/context-guard),
remote config, start, stop, drain-inbox, restart, and status handlers into
a dedicated tryHandleAgentsProcess function.

INBOX_DRAIN_CAP moves with its only consumer (drain-inbox handler).
tryHandleAgentsProcess is called after models, before team/crud (order preserved).

agents.ts: 1612 -> 1444 lines (-168) | agents-process.ts: 199 lines new
